Move shim runtimes to shim files
