+++ src/pushStateTree.js
@@ -2,46 +2,21 @@ var root = typeof window !== 'undefined' && window || global;
 var document = root.document;
 var location = root.location;
 
-var isIE = (function(){
-  var trident = window.navigator.userAgent.indexOf('Trident');
-  return trident >= 0;
-}());
+// Add support to location.origin for all browsers
+require('./origin.shim');
 
-// Shim, to work with older browsers
-(function () {
-  // Opera and IE doesn't implement location.origin
-  if (!root.location.origin) {
-    root.location.origin = root.location.protocol + '//' + root.location.host;
-  }
-})();
+let isIE = require('./ieOld.shim').isIE;
 
-(function () {
-  /* global HTMLDocument */
-  if (Function.prototype.bind) { return; }
-
-  Function.prototype.bind = function (oThis) {
-    if (typeof this !== 'function') {
-      // closest thing possible to the ECMAScript 5 internal IsCallable function
-      throw new TypeError('Function.prototype.bind - what is trying to be bound is not callable');
-    }
-
-    var aArgs = Array.prototype.slice.call(arguments, 1),
-      fToBind = this,
-      FNOP = function () {},
-      fBound = function () {
-        var context = oThis;
-        if (this instanceof FNOP && oThis){
-          context = this;
-        }
-        return fToBind.apply(context, aArgs.concat(Array.prototype.slice.call(arguments)));
-      };
-
-    FNOP.prototype = this.prototype;
-    fBound.prototype = new FNOP();
-
-    return fBound;
-  };
-})();
+// If you have your own shim for ES3 and old IE browsers, you can remove this shim from your package by adding a
+// webpack.DefinePlugin that translates `typeof PST_NO_OLD_IE === 'undefined'` to false, this will remove the section
+// in the minified version:
+//     new webpack.DefinePlugin({
+//       PST_NO_OLD_IE: false
+//     })
+// https://webpack.github.io/docs/list-of-plugins.html#defineplugin
+if (typeof PST_NO_OLD_IE === 'undefined') {
+  require('./es3.shim.js');
+}
 
 // IE9 shims
 var HashChangeEvent = root.HashChangeEvent;
@@ -63,16 +38,16 @@ var Event = root.Event;
 
   CustomEvent.prototype = Event.prototype;
 
-  if (!root.CustomEvent || !!isIE) {
+  if (!root.CustomEvent || isIE()) {
     root.CustomEvent = CustomEvent;
   }
 
   // Opera before 15 has HashChangeEvent but throw a DOM Implement error
-  if (!HashChangeEvent || (root.opera && root.opera.version() < 15) || !!isIE) {
+  if (!HashChangeEvent || (root.opera && root.opera.version() < 15) || isIE()) {
     HashChangeEvent = root.CustomEvent;
   }
 
-  if (isIE) {
+  if (isIE()) {
     Event = CustomEvent;
   }
 
@@ -229,71 +204,6 @@ var Event = root.Event;
   HTMLDocument.prototype.dispatchEvent = dispatchEventFunc;
   Element.prototype.dispatchEvent = dispatchEventFunc;
   window.dispatchEvent = dispatchEventFunc;
-})();
-
-(function () {
-  // modern browser support forEach, probably will be IE8
-  var modernBrowser = 'forEach' in Array.prototype;
-
-  // IE8 pollyfills:
-  // IE8 slice doesn't work with NodeList
-  if (!modernBrowser) {
-    var builtinSlice = Array.prototype.slice;
-    Array.prototype.slice = function() {
-      var arr = [];
-      for (var i = 0, n = this.length; i < n; i++) {
-        if (i in this) {
-          arr.push(this[i]);
-        }
-      }
-
-      return builtinSlice.apply(arr, arguments);
-    };
-  }
-  if (!('forEach' in Array.prototype)) {
-    Array.prototype.forEach = function(action, that) {
-      for (var i = 0; i < this.length; i++) {
-        if (i in this) {
-          action.call(that, this[i], i);
-        }
-      }
-    };
-  }
-  if (typeof String.prototype.trim !== 'function') {
-    String.prototype.trim = function() {
-      return this.replace(/^\s+|\s+$/g, '');
-    };
-  }
-  if (!Array.prototype.filter) {
-    Array.prototype.filter = function(fun /*, thisArg */) {
-      if (this === void 0 || this === null) {
-        throw new TypeError();
-      }
-
-      var t = Object(this);
-      var len = t.length >>> 0;
-      if (typeof fun !== 'function') {
-        throw new TypeError();
-      }
-
-      var res = [];
-      var thisArg = arguments.length >= 2 ? arguments[1] : void 0;
-      for (var i = 0; i < len; i++) {
-        if (i in t) {
-          var val = t[i];
-
-          // NOTE: Technically this should Object.defineProperty at
-          //       the next index, as push can be affected by
-          //       properties on Object.prototype and Array.prototype.
-          //       But that method's new, and collisions should be
-          //       rare, so use the more-compatible alternative.
-          if (fun.call(thisArg, val, i, t)) { res.push(val); }
-        }
-      }
-
-      return res;
-    };
-  }
 })();
 
 // Constants for uglifiers
@@ -514,9 +424,9 @@ function PushStateTree(options) {
 
     // If there is holding dispatch in the event, do it now
     if (holdingDispatch) {
-      this.dispatch();
+      rootElement.dispatch();
     }
-  }.bind(rootElement));
+  });
 
   var readOnhashchange = false;
   var onhashchange = function () {
@@ -535,9 +445,9 @@ function PushStateTree(options) {
 
     // If there is holding dispatch in the event, do it now
     if (holdingDispatch) {
-      this.dispatch();
+      rootElement.dispatch();
     }
-  }.bind(rootElement);
+  };
 
   rootElement.avoidHashchangeHandler = function () {
     // Avoid triggering hashchange event
@@ -560,7 +470,7 @@ function PushStateTree(options) {
   root.addEventListener('load', function () {
     dispatchHashChange();
 
-    if (isIE) {
+    if (isIE()) {
       root.setInterval(() => {
         if (this.uri !== oldURI) {
           dispatchHashChange();
